se realizo varios cambios

+++ src/main/java/com/izenshy/gessainvoice/modules/product/stock/dto/StockDTO.java
@@ -8,8 +8,8 @@ public class StockDTO {
     private Long outletId;
     private float stockQuantity;
     private Boolean stockAvalible;
-    private float unitPrice;
-    private float pvpPrice;
+    private double unitPrice;
+    private double pvpPrice;
     private int stockMax;
     private int stockMin;
     private Boolean applyTax;
