+++ src/main/java/com/izenshy/gessainvoice/modules/product/stock/dto/StockDeluxeDTO.java
@@ -13,8 +13,8 @@ public class StockDeluxeDTO {
     private Long outletId;
     private float stockQuantity;
     private Boolean stockAvalible;
-    private float unitPrice;
-    private float pvpPrice;
+    private double unitPrice;
+    private double pvpPrice;
     private int stockMax;
     private int stockMin;
     private Boolean applyTax;
